### 1.2.2
- Refactored: Split widgets and components folders.
- Updated: the mobile handling of the contact section.
