+++ lib/components/components.dart
@@ -2,5 +2,4 @@ export "desktop/desktop.dart";
 export "mobile/mobile.dart";
 export "menu/menu.dart";
 export "projects/projects.dart";
-export "contact/content_contact.dart";
 export "misc/not_found.dart";
