+++ lib/components/contact/contact.dart
@@ -1,0 +1,3 @@
+export "contact_column.dart";
+export "contact_field.dart";
+export "contact_title.dart";
