+++ lib/components/desktop/desktop_home.dart
@@ -2,7 +2,7 @@ import "package:flutter/material.dart";
 import "package:get/get.dart";
 import "package:x_containers/x_containers.dart";
 
-import "../../widgets.dart";
+import "../../widgets/widgets.dart";
 
 /// The content of the body in the "Home" mode.
 class BodyContentsHome extends StatelessWidget {
