+++ lib/components/menu/desktop_menu.dart
@@ -1,10 +1,7 @@
 import "package:flutter/material.dart";
 import "package:x_containers/x_containers.dart";
 
-import "desktop_title.dart";
-import "cookies_card.dart";
-import "menu_list.dart";
-import "test_card.dart";
+import "menu.dart";
 
 /// The half of the body containing the title and the menu.
 class BodyMenu extends StatelessWidget {
