--- lib/widgets/body/desktop/desktop.dart
@@ -1,3 +1,0 @@
-export "desktop_home.dart";
-export "desktop_projects.dart";
-export "desktop_contents.dart";
